fix token group heading, many token sets issue
